requirements.txt
flask==0.12.2
flask-security==3.0.0
flask-sqlalchemy==2.4.1
flask_restful
flask-login==0.3.2
pandas==0.23.4
xlrd==1.1.0
redis==3.5.0
flask_migrate==2.5.3
flask_cors==3.0.8
flask_script==2.0.6
flask-shell-ipython==0.3.1
PyMySQL==0.9.3
celery==4.4.2
flask-marshmallow==0.12.0
marshmallow-sqlalchemy==0.23.0
numpy==1.18.3
scipy==1.4.1
flower==0.9.4
black==19.10b0
bcrypt==3.1.7
email-validator==1.1.1
